Actulizacion de plugin 2016

+++ wp-content/plugins/wp-statistics/wps-install.php
@@ -41,13 +41,33 @@
 			hits int(11),
 			honeypot int(11),
 			PRIMARY KEY  (ID),
-			UNIQUE KEY date_ip_agent (last_counter,ip,agent (75),platform (75),version (75)),
+			UNIQUE KEY date_ip_agent (last_counter,ip,agent(75),platform(75),version(75)),
 			KEY agent (agent),
 			KEY platform (platform),
 			KEY version (version),
 			KEY location (location)
 		) CHARSET=utf8");
 		
+		$create_visitor_table_old = ("CREATE TABLE {$wp_prefix}statistics_visitor (
+			ID int(11) NOT NULL AUTO_INCREMENT,
+			last_counter date NOT NULL,
+			referred text NOT NULL,
+			agent varchar(255) NOT NULL,
+			platform varchar(255),
+			version varchar(255),
+			UAString varchar(255),
+			ip varchar(60) NOT NULL,
+			location varchar(10),
+			hits int(11),
+			honeypot int(11),
+			PRIMARY KEY  (ID),
+			UNIQUE KEY date_ip_agent (last_counter,ip,agent (75),platform (75),version (75)),
+			KEY agent (agent),
+			KEY platform (platform),
+			KEY version (version),
+			KEY location (location)
+		) CHARSET=utf8");
+
 		$create_exclusion_table = ("CREATE TABLE {$wp_prefix}statistics_exclusions (
 			ID int(11) NOT NULL AUTO_INCREMENT,
 			date date NOT NULL,
@@ -81,6 +101,19 @@
 			UNIQUE KEY uri (uri)
 		) CHARSET=utf8");
 		
+		$create_search_table = ("CREATE TABLE {$wp_prefix}statistics_search (
+			ID bigint(20) NOT NULL AUTO_INCREMENT,
+			last_counter date NOT NULL,
+			engine varchar(64) NOT NULL,
+			host varchar(255),
+			words varchar(255),
+			visitor bigint(20),
+			PRIMARY KEY  (ID),
+			KEY last_counter (last_counter),
+			KEY engine (engine),
+			KEY host (host)
+		) CHARSET=utf8");
+
 		// Before we update the historical table, check to see if it exists with the old keys
 		$result = $wpdb->query( "SHOW COLUMNS FROM {$wp_prefix}statistics_historical LIKE 'key'" );
 		
@@ -100,12 +133,57 @@
 		dbDelta($create_exclusion_table);
 		dbDelta($create_pages_table);
 		dbDelta($create_historical_table);
+		dbDelta($create_search_table);
+		
+		// Some old versions (in the 5.0.x line) of MySQL have issue with the compound index on the visitor table
+		// so let's make sure it was created, if not, use the older format to create the table manually instead of
+		// using the dbDelta() call.
+		$dbname = DB_NAME;
+		$result = $wpdb->query("SHOW TABLES WHERE `Tables_in_{$dbname}` = '{$wpdb->prefix}statistics_visitor'" );
 
-		$wpdb->query( "DROP INDEX `date_ip` ON {$wp_prefix}statistics_visitor" );
+		if( $result != 1 ) {
+			$wpdb->query( $create_visitor_table_old );
+		}
+		
+		// Check to see if the date_ip index still exists, if so get rid of it.
+		$result = $wpdb->query("SHOW INDEX FROM {$wp_prefix}statistics_visitor WHERE Key_name = 'date_ip'");
+
+		// Note, the result will be the number of fields contained in the index.
+		if( $result > 1 ) {
+			$wpdb->query( "DROP INDEX `date_ip` ON {$wp_prefix}statistics_visitor" );
+		}
+		
+		// One final database change, drop the 'AString' column from visitors if it exists as it's a typo from an old version.
+		$result = $wpdb->query( "SHOW COLUMNS FROM {$wp_prefix}statistics_visitor LIKE 'AString'" );
+		
+		if( $result > 0 ) {
+			$wpdb->query( "ALTER TABLE `{$wp_prefix}statistics_visitor` DROP `AString`" );
+		}
 		
 		// Store the new version information.
 		update_option('wp_statistics_plugin_version', WP_STATISTICS_VERSION);
 		update_option('wp_statistics_db_version', WP_STATISTICS_VERSION);
+
+		// Now check to see what database updates may be required and record them for a user notice later.
+		$dbupdates = array( 'date_ip_agent' => false, 'unique_date' => false );
+		
+		// Check the number of index's on the visitors table, if it's only 5 we need to check for duplicate entries and remove them
+		$result = $wpdb->query("SHOW INDEX FROM {$wp_prefix}statistics_visitor WHERE Key_name = 'date_ip_agent'");
+
+		// Note, the result will be the number of fields contained in the index, so in our case 5.
+		if( $result != 5 ) {
+			$dbupdates['date_ip_agent'] = true;
+		}
+		
+		// Check the number of index's on the visits table, if it's only 5 we need to check for duplicate entries and remove them
+		$result = $wpdb->query("SHOW INDEX FROM {$wp_prefix}statistics_visit WHERE Key_name = 'unique_date'");
+
+		// Note, the result will be the number of fields contained in the index, so in our case 1.
+		if( $result != 1 ) {
+			$dbupdates['unique_date'] = true;
+		}
+
+		$WP_Statistics->update_option( 'pending_db_updates', $dbupdates );
 
 		// Get the robots list, we'll use this for both upgrades and new installs.
 		include_once('robotslist.php');
@@ -115,6 +193,9 @@
 			// If this is a first time install, we just need to setup the primary values in the tables.
 		
 			$WP_Statistics->Primary_Values();
+			
+			// By default, on new installs, use the new search table.
+			$WP_Statistics->update_option('search_converted', 1);
 			
 		} else {
 
@@ -176,6 +257,20 @@
 			// However that seems VERY unlikely.
 			$exclude_admins = $WP_Statistics->get_option('exclude_administrator', '2');
 			if( $exclude_admins == '2' ) { $WP_Statistics->update_option('exclude_administrator', '1'); }
+			
+			// WordPress 4.3 broke the diplay of the sidebar widget because it no longer accepted a null value
+			// to be returned from the widget update function, let's look to see if we need to update any 
+			// occurances in the options table.
+			$widget_options = get_option( 'widget_wpstatistics_widget' );
+			if( is_array( $widget_options ) ) {
+				foreach( $widget_options as $key => $value ) {
+					// We want to update all null array keys that are integers.
+					if( $value === null && is_int( $key ) ) { $widget_options[$key] = array(); }
+				}
+
+				// Store the widget options back to the database.
+				update_option( 'widget_wpstatistics_widget', $widget_options );
+			}
 		}
 
 		// If this is a first time install or an upgrade and we've added options, set some intelligent defaults.
@@ -197,6 +292,7 @@
 		if( $WP_Statistics->get_option('robotlist') === FALSE ) { $WP_Statistics->store_option('robotlist',$wps_robotslist); }
 		if( $WP_Statistics->get_option('exclude_administrator') === FALSE ) { $WP_Statistics->store_option('exclude_administrator',TRUE); }
 		if( $WP_Statistics->get_option('disable_se_clearch') === FALSE ) { $WP_Statistics->store_option('disable_se_clearch',TRUE); }
+		if( $WP_Statistics->get_option('disable_se_ask') === FALSE ) { $WP_Statistics->store_option('disable_se_ask',TRUE); }
 		if( $WP_Statistics->get_option('map_type') === FALSE ) { $WP_Statistics->store_option('map_type','jqvmap'); }
 
 		if( $WPS_Installed == false ) {
